updated SIGQUIT handling and seperated signalhandling into its on file
