+++ first_try/main.c
@@ -6,14 +6,11 @@
 /*   By: tvalimak <Tvalimak@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 11:16:09 by akuburas          #+#    #+#             */
-/*   Updated: 2024/04/04 18:45:20 by akuburas         ###   ########.fr       */
+/*   Updated: 2024/04/08 18:38:05 by tvalimak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include <readline/history.h>
-#include <readline/readline.h>
-#include <termios.h>
 
 // when in heredoc and CTR+C the ^C will not display
 // when in bash and you have inputted cat for example, then you CTRL+C the ^C will display
@@ -21,46 +18,6 @@
 // When terminal is opened for the first time and you do CTRL+C in bash, the ^C is shown for once,
 // need to figure out why.
 // CTRL + \ = SIGQUIT
-
-static void	signal_handler(int signal)
-{
-	if (signal == CTRL_C)
-	{
-		write(1, "\n", 1);
-		rl_on_new_line();
-		rl_replace_line("", 0);
-		rl_redisplay();
-	}
-}
-
-static void	carrot_toggle(int on)
-{
-	struct termios	term;
-
-	tcgetattr(STDIN_FILENO, &term);
-	if (!on)
-		term.c_lflag &= ~ECHOCTL;
-	else
-		term.c_lflag |= ECHOCTL;
-	tcsetattr(STDIN_FILENO, TCSANOW, &term);
-}
-
-static void	set_state(t_state state)
-{
-	if (state == DEFAULT)
-	{
-		carrot_toggle(1);
-	}
-	if (state == HEREDOC)
-	{
-		return ;
-	}
-	if (state == HANDLER)
-	{
-		carrot_toggle(0);
-	}
-}
-
 
 int	duplicate_env(char **env, t_shelldata *data)
 {
@@ -111,9 +68,9 @@ int	main(int argc, char **argv, char **env)
 	{
 		set_state(HANDLER);
 		//set_state(DEFAULT);
-		signal(CTRL_C, signal_handler);
-		input = readline("bananashell-0.05:");
-		input = readline("bananashell-0.07:");
+		signal(SIGINT, signal_handler);
+		signal(SIGQUIT, signal_handler);
+		input = readline("bananashell-0.09:");
 		if (!input)
 		{
 			printf("exit\n");
