+++ first_try/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tvalimak <Tvalimak@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 11:16:05 by akuburas          #+#    #+#             */
-/*   Updated: 2024/04/03 12:52:22 by tvalimak         ###   ########.fr       */
+/*   Updated: 2024/04/08 18:41:09 by tvalimak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,6 +31,9 @@
 
 /*libft functions*/
 # include "libft/libft.h"
+
+/*for termios*/
+#include <termios.h>
 
 /*Definition for SIGINT*/
 # define CTRL_C SIGINT
@@ -115,6 +118,10 @@ int			rl_clear_history(void);
 int			mini_split(char *input, t_shelldata *data);
 void		free_double_array(char ***array);
 void		rl_replace_line(char *str, int num);
-// static void	signal_handler(int signal);
+int			parse_split_input(t_shelldata *data);
+int			split_cleaner(t_shelldata *data);
+void		signal_handler(int signal);
+void		carrot_toggle(int on);
+void		set_state(t_state state);
 
 #endif
