+++ first_try/signal.c
@@ -1,0 +1,56 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tvalimak <Tvalimak@student.hive.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/08 18:37:08 by tvalimak          #+#    #+#             */
+/*   Updated: 2024/04/08 18:40:27 by tvalimak         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	signal_handler(int signal)
+{
+	if (signal == SIGINT)
+	{
+		write(1, "\n", 1);
+		rl_on_new_line();
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}
+	if (signal == SIGQUIT)
+	{
+		rl_redisplay();
+	}
+}
+
+void	carrot_toggle(int on)
+{
+	struct termios	term;
+
+	tcgetattr(STDIN_FILENO, &term);
+	if (!on)
+		term.c_lflag &= ~ECHOCTL;
+	else
+		term.c_lflag |= ECHOCTL;
+	tcsetattr(STDIN_FILENO, TCSANOW, &term);
+}
+
+void	set_state(t_state state)
+{
+	if (state == DEFAULT)
+	{
+		carrot_toggle(1);
+	}
+	if (state == HEREDOC)
+	{
+		return ;
+	}
+	if (state == HANDLER)
+	{
+		carrot_toggle(0);
+	}
+}
